tested and minor fixes to reg_score
